requirements.txt
tensorflow==2.15.0
tensorflow-model-optimization==0.8.0
kagglehub>=0.4.1
numpy==1.26.4
pandas==2.1.4
Pillow==10.2.0
pyyaml==6.0.1
tqdm==4.66.2
